Lesson_3/src/ru/geelbrains/datastructure/ads/MyHW/DequeImpl.java
package ru.geelbrains.datastructure.ads.MyHW;

import ru.geelbrains.datastructure.ads.queue.QueueImpl;

public class DequeImpl<E> extends QueueImpl<E> implements Deque<E> {

    public DequeImpl(int maxSize) {
        super(maxSize);
    }

    @Override
    public boolean insertLeft(E value) {
        if (isFull()) {
            return false;
        }

        if (head - 1 < 0){
            head = data.length;
        }

        data[--head] = value;
        size++;
        return true;
    }

    @Override
    public boolean insertRight(E value) {
        return super.insert(value);
    }

    @Override
    public E removeLeft() {
        return super.remove();
    }

    @Override
    public E removeRight() {
        if (isEmpty()) {
            return null;
        }

        if (tail < 0) {
            tail = data.length - 1;
        }

        E removedValue = data[tail--];
        size--;
        return removedValue;

    }
}
